+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [x.x.x] - 26-Apr-2020
+- Added Terraform provider for Grafana
+- Automated the upload of dashboards to Grafana
+- Automated the creation of datasource
+- Added gitignore file
+
 ## [x.x.x] - 13-Apr-2020
 - Added Dockerfile for automatic Grafana DataSource & Dashboards initialisation
 - Added Readme for docker